feat: Adding electron builds for Mac, Win, and Linux BREAKING CHANGE: 'start' and 'electron-dev' now use port 3000 (was non-standard 3001) fix: Loading of assets in non-root serves/hosting (header sets ) fix: Redirect to app relative 'home' if loading app from "index.html" (catch-all in router)
Windows exe, MacOS dmg, Linux (deb,rpm). AppImage and Snap are easy additionas as well.

The routing was changed to have a 'catch-all' to route to '/' due to how
initial page loads work with electron.

A bash script was added to use docker for building windows exe on linux.
This is because otherwise you need Wine installed on your system, which
is a very large dependency.